requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.9.24
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
Django==4.1.4
django-environ==0.9.0
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
filelock==3.9.0
flake8==5.0.4
flake8-django==1.1.5
geographiclib==2.0
geopy==2.3.0
identify==2.5.11
isort==5.11.4
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.3
platformdirs==2.6.0
pre-commit==2.21.0
prompt-toolkit==3.0.33
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pygal==3.0.0
pygal-maps-world==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
uWSGI==2.0.21
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.5
